Feat: settings WIP
Added a method to load the settings.json file and supply it to the methods that need it
added a reminder to add passwords and to re-check the old manage ip list function.
added a placeholder toggle_lists route that was missing.
the next step is to add have each of the 3 functions mutate the json as required and feed it into a shared "write  settings" function, just need to do some debuging first.