Chore: list function tidy up, response follow up prep
Moved the js for populating the management selects to it's own file, both for separation of concerns (achieving that on this project is going to be a pain) and because it will also be called by Call_endpoints to updated those selects when it makes changes to files.
Modified list_books to return data in the format that the management selects recognise, gave it a properly formatted response and changed the generic name "response" to "list".
Added 2 optional parameters to Call_and_display, they are Update_lib_selects and show_ip_lists, both false by default.
These settings control updating the management selects when an api call updates the file structure, and showing the IP white/blacklists respectively, since that info is not currently visible to the user (dont wory, its still behind a password).
Started work on the first of these options, tommorow's work will likely revolve around finishing these 2 functions and modifying the calls to the function to consider the new options.