Feat: Auto update management selects
Added a re-usable import for Populate_management_selects to call_endpoints so that it can automaticaly update the selects the users will be relying on in the management page.
Added removal of old select content to populate_managment_selects to stop it pilling up under repeated calls.
Removed the export from Pop_management_selects, since it is not a module, replaced it with a standard call.
Added the necessary optional paramaters to use this functionality on several of the API call handlers.
Updated todo to reflect progress.