Feat: clear thumbs endpoint
Added an endpoint at /clear-thumbs to empty the thumbnail cache, with the option to re-populate the cache and remove manualy uploaded files.
Removed unneeded comment in populate_management_select.
updated todo
reformatted Management.php to have new inputs to match the endpoint.
added event handlers for the new endpoint and it's form