Feat: secured json writes, extra responses
Added password checks to the fetch/write settings functions.
Modified Call_and_display with a code parameter, since functions that require setting now need a password, modified the appropriate calls to this function to pass it.
Added a few extra success response messages so more codes can be used.
Added functions for reading unsecured json files.
Modified Call_and_display to only show the IP lists if the user was authorised, since that's not the sort of thing we want unathorised people to see, this works because the functions with show_ip_lists set to true only respond with 200 or errors