Feat: Reassign thumb
Added and endpoint at /reassign-thumb that allows the user to replace the thumbnails assigned to books.
Added some example/ placeholder data to thumbnail_map.json since the code to populate it is not ready yet.
Added some notes on the thumbnail system to the TODO.
Fixed missing capitalisation and a <p> that should have been a label, gave thumbnail select a more consistent name in Management.php