Refactor: updated append_H&F
Updated the ancient ajax based append_H&F I have been using for years to a more modern jquery approach.
Created a copy of the old version in the depreciated folder.
Fixed redundant import of jquery in Header.html, since if the header has loaded we know for a fact that we already have jquery.
Updated todo.txt