Feat: json-book list endpoint
Added an endpoint at /json-list-books that returns the book list in a json format, rather than html.
Added call to this new endpoint and the json map to pop_library entries.
This is the first step in refactoring this page and the management page to use json data rather than the old "read what we left in a div" approach that was used by the php.
Note: In all likelihood this endpoint will eventually replace the existing /list-books endpoint, most likely taking it's name and moving the old endpoint to /Docs/Depreciated.