Refactor: convert thumb_library to use json
Added a new endpoint at /list-thumbs-json to return a json list of the thumbnails, similar to the last 2 endpoints this will likely replace the older versions that imitate the php version by returning html.
The endpoint /thumb-map now returns the json thumb list as a second object.
Updated todo.txt