Fix: fixed thumb selects, tested selects and endpoints
Investigated remaining issues: fixed use of in on a for loop, the select weirdness was because the script was calling select from within select's own handler, using a this is a far neater, more sensible solution.
Added an "API not available" message to management.html and disabled it should there be no connection.
Updated  pop library entries "No API" message to use the same language as the management one.
The following are issues noticed and resolved during the testing:
The upload success pages where linking to the old management page, fixed.
Incorrect capitalisation of a folder name in 403 response.

Updated todo