Feat: Library dropdowns
Used the same technique as in the management page to add dropdowns to each library folder.
Now only the title is visible by default, with the list becoming visible when the title is pressed. The title box changes background colour and the "expand more" arrow flips to indicate the change.
Added style for the new decorated folder titles
Renamed management-dropdown to management-import to be more correct.
Renamed Included_public_domain to included public domain
Updated TODO