Fix: Pop_managment_select erroring, epub reader test moved
Fixed a bug in pop_management select where the check that a book select that matches a parent select exists before manipulating it would always return true since it was checking if the result list was undefined (which it never is) instead of the list value itself.
Updated the missing spaces message in the library to reflect the removal of the list of offending books (the list was long and ugly since it intentional displayed poorly formatted text).
moved epub_embed_test from depreciated to pages, imported epub.js - still WIP.
Re-organised and updated TODO.txt