Feat: Auto updating thumb lists, clearing text inputs
Added a function to the thumbnail management API calls to update the thumb lists themselves right after calling, will probably transition the book selects to this system to as it's easier to follow and cuts down on the number of listeners.
Set the management listeners to automatically remove the contents of text inputs to save the user time and hassle (and to not leave a passwords on screen).
Updated TODO.txt