Feat: Server IP source of truth, IP restrictions, set target IP
Added  Server_IP.js, a script that sets a cookie containing the server's local address.
Updated populate_library entries to use the cookie for it's addresses and to point the user to management if the api is unreachable.
updated pop_management_selects with the new addressing system and a form to set the cookie when the API is unreachable.
Updates set_management_listeners.js to use the same addressing as above.
Added a rule to BookAPI.py to reject any requests from non-local addresses.
Added some extra response codes to call_and_display for 421 "Misdirected" (for non-local IPs) and 451 "legal reasons" because it was mildly amusing.
Imported the new set_ip script to each of the main pages to make sure they have access.
Did a quick pass of adding some missing lets to variables.
Removed  debug console log from populate_support_table.js
Updated TODO.txt