Refactor: getdata and pop management select module / update script.
Fixed the issue with calling populate_managment selects as a module in call_and_display (if called as a module it wouldn't run, if not a module it would error on the imports) by creating a specific update_management_selects script, this has the bonus of keeping the module and not having to run unnecesary startup code on each update.
Added a get_data script so that functions that need JSON and not http/html can all use the same function (Hooking up intances WIP).
made populate_management_selects a module, letting it import get_data and set_server_address.
Updated TODO, moved the cookie update because I missed that last time.