Feat: sitemap & matching endpoint
Added a sitemap in both xml and txt formats.
Added an endpoint to display the sitemap on request.
Added a redirect so that the management page reloads after setting a server IP so that the user doesn't have to (came up when getting someone else to test yesterday's update).
Wrapped the homepage instruction segments in links to the pages they talk about (this also came up in yesterday's user test)
Updated todo.txt