Feat: delete thumbnail endpoint
Added an endpoint at /delete-thumb with the parameter: target
Added the matching form and help entry.
Started working on populating the cache, currently investigating trying to extract/find ISBNs vs using the name and hoping google books catches it, have confirmed that the books api can return the required data and image.
(the above research is when the need for an individual delete endpoint became apparent, because entering names was not very reliable)