Refactor: API segmentation
Experiment in segmenting the API, it is over 700 lines at time of writing and since  importing is such a big part of python I have decided to separate out some of the functions.
Experimented with saving the updated book data json, looks like it will require some interpreting.
Added note to readme.md