Feat: edit book data endpoint
Added an endpoint at /edit-book-data that allows users to modify the data of a given book and validate it to make the data generator ignore it.
Added an endpoint at /get-book-data, that returns the book data json, should be useful for making this a single source of truth for book data and help replace the thumb map json.
Added form to populate the new data (will hook up to an event listener with validation)
Updated TODO