Feat: edit book data UI
Hooked up the edit book data form to it's endpoint via an event handler (debugging one with so many variables took forever)
Added validation and a few safeties to the edit_book_Data endpoint.
Working on setting up book data as a single source of truth for the library, management selects and thumbnails, debating between importing it everywhere or using it to update existing sources like thumb_map.