Refactor: Auto validate user edited book data (& fix: upload of gen data)
Modified the edit book data form and it's handler to always send the data with the validated attribute set to "True" to prevent any user inputted data from being overwritten by the function.
Fixed the issue that was stopping gen_book_data from writing correctly, so yesterday's improvements are being uploaded.
Added handling for book_data's thumbnail attribute, since it uses paths and needs "/", which flask doesnt like (converts the / to -@ when sent and back when processed).
Added an easter egg for the http 418 response code.