Refactor: delete thumb using book data
Updated the delete thumb endpoint code to use the new folder structure and update book data (hopefully this should be faster for the other functions, now that the process is understood).
Started setting up UI for the updated endpoint, needs update to setup new option selects, similar to the way book selects work with the library.