Fix: move book saving
Fixed an issue where delete_book couldn't load the book_Data json if called from another file since the file would already be open (and likely remain in use since the caller itself will want to save).
Added an optional parameter book_data to BD_delete_book to allow other book data functions that call it to provide it with book data without it having to open or rely on the existing bookdata  file (useful if it's already open or has been modified.