Feat: moved book tagging
Fixed the system for tagging books that have been moved from a deleted folder and added said tagging system delete folder's book_data function.
Added a safety check to delete_folder to ensure that there is a safe spot for all of it's books before anything is deleted.
Added a system to allow books from a deleted folder to be moved to uploads if there is already a book with the same name in misc (integrating this with the other 2 systems took surprisingly long).
Updated the HTTP 409 response to include a response message for conflicted message.
Added code to the BD_delete_thumb cache to re-format the thumbnail data (This entire endpoint might be deleted since it doesn't add anything and can destroy alot).