Refactor: Populate_library_entries using Book_data.json
Updated the library populating script to use the new data format and made some improvements/ bugfixes.
Reduced the number of API calls from 2 to 1 using a call to get-book-and-thumb-data rather than the separate thumb and book endpoints (depreciated and based on the old approach respectively).
Moved reading of File_type_support.json out of the book creating loop so that it is only opened once.
Many reference fixes and name improvements (such as list of BOOK_MAP.Books -> book_folder of LIB_DATA.Books.folder)
Changed the thumbnail source assignment to directly take the book's extension value (simpler and more reliable than the old method), added extra checks for the thumbnail and set an onerror and an else condition to fetch a placeholder thumb if the original is invalid (made the switch a function for this reason).
Updated TODO.txt